+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Add Letsencrypt support using the `greenlock` module
 * Added websocket proxy support
+* Ports are tested before a node process starts
+* Should a port still be in use when a new process starts hohenheim will try a new one
 
 ## 0.1.0